# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для LMS платформы Cap Education.",
        "title": "Cap Education LMS - API",
        "contact": {},
        "version": "1.0"
    },
    "host": "capapi.grtsq.ru",
    "basePath": "/",
    "paths": {
...
        "/profile": {
            "get": {
                "description": "Получить данные текущего авторизованного пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "USER: Мой профиль",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменение личных данных и загрузка аватара.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "USER: Обновить профиль",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Фото профиля",
                        "name": "avatar",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Имя",
                        "name": "first_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Фамилия",
                        "name": "last_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Телефон",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Населенный пункт",
                        "name": "city",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Родной язык",
                        "name": "language",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Учебное заведение",
                        "name": "school_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "WhatsApp ссылка",
                        "name": "whatsapp",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Telegram ссылка",
                        "name": "telegram",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/profile/teacher/schedule": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "USER: Сохранить график работы учителя",
                "parameters": [
                    {
                        "description": "JSON расписания (любой формат Игоря)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/staff/submissions/{id}/evaluate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff-Review"
                ],
                "summary": "STAFF: Проверить ДЗ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Результат проверки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.EvaluateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Role": {
            "type": "string",
            "enum": [
                "student",
                "parent",
                "teacher",
                "moderator",
                "curator",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleParent",
                "RoleTeacher",
                "RoleModerator",
                "RoleCurator",
                "RoleAdmin"
            ]
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/domain.Role"
                },
                "school_name": {
                    "type": "string"
                },
                "telegram": {
                    "type": "string"
                },
                "whatsapp": {
                    "type": "string"
                }
            }
        },
...
        "http.EvaluateRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "grade": {
                    "type": "integer"
                },
                "is_accepted": {
                    "type": "boolean"
                }
            }
        },
...
    }
}